requirements.txt
# Core ML/AI Libraries
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0

# Labelbox Integration
labelbox>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# File I/O and Data Formats
ndjson>=0.3.1
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# Logging and Utilities
tqdm>=4.65.0
pathlib2>=2.3.7

# Optional: For advanced video processing
# imageio>=2.31.0
# imageio-ffmpeg>=0.4.8

# Development and Testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0 